Make routes simpler and URLs more user-friendly
Was "/filesystem" and "/filesystem_details/1",
now "/filesystems" and "/filesystems/1".

It is supposed to make easier navigation (e. g. by removing the filesystem id
a user will get to the list of all filesystems), and reasoning about the code.
