--- iml-gui/crate/src/page/volume_detail.rs
@@ -1,10 +1,0 @@
-use crate::Msg;
-use seed::prelude::*;
-
-pub struct Model {
-    pub id: u32,
-}
-
-pub fn view(_model: &Model) -> impl View<Msg> {
-    seed::empty()
-}
