+++ iml-gui/crate/src/page/volumes.rs
@@ -1,9 +1,5 @@
-use crate::Msg;
+use crate::{Model, Msg};
 use seed::prelude::*;
-
-pub struct Model {
-    pub id: u32,
-}
 
 pub fn view(_model: &Model) -> impl View<Msg> {
     seed::empty()
